pyproject.toml
[tool.poetry]
name = "donate4.fun"
version = "0.1.0"
description = ""
authors = ["Nikolay Bryskin <nbryskin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.22.0"
fastapi = "^0.77.1"
qrcode = "^7.3.1"
uvicorn = "^0.17.6"
Jinja2 = "^3.1.2"
pymaging = {git = "https://github.com/ojii/pymaging.git"}
pymaging-png = {git = "https://github.com/ojii/pymaging-png.git"}
python-multipart = "^0.0.5"
Authlib = "<1.0"
PyYAML = "^6.0"
itsdangerous = "^2.1.2"
jwt = "^1.3.1"
aiogoogle = "^4.1.0"
databases = {extras = ["asyncpg"], version = "^0.5.5"}
asyncpg = "^0.25.0"
hypercorn = "^0.13.2"
email-validator = "^1.2.1"
fastapi-debug-toolbar = "^0.2.1"
dataclasses-json = "^0.5.7"
PyJWT = "^2.4.0"
Mako = "^1.2.0"
starlette-authlib = "^0.1.9"
multiavatar = "^1.0.3"
funkybob = {git = "https://github.com/nikicat/funkybob"}
setuptools = "^62.3.2"

[tool.poetry.dev-dependencies]
ipython = "^8.3.0"
flake8 = "^4.0.1"
pytest = "^7.1.2"
SQLAlchemy-Utils = "^0.38.2"
anyio = "^3.6.1"
pytest-icdiff = "^0.5"
vcrpy = "^4.1.1"
pytest-vcr = "^1.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
